Move plugin system to core library
