+++ src/core/plugins/plugin.h
