+++ src/core/plugins/plugininfo.cpp
