+++ src/core/plugins/plugininfo.h
