+++ src/core/plugins/pluginmanager.cpp
