+++ src/core/plugins/pluginmanager.h
