+++ src/gui/library/coverwidget.cpp
@@ -22,7 +22,7 @@
 #include <core/library/musiclibrary.h>
 #include <core/models/track.h>
 #include <core/player/playermanager.h>
-#include <pluginsystem/pluginmanager.h>
+#include <core/plugins/pluginmanager.h>
 
 #include <QHBoxLayout>
 #include <QLabel>
