+++ src/gui/library/statuswidget.cpp
@@ -23,7 +23,7 @@
 
 #include <core/models/track.h>
 #include <core/player/playermanager.h>
-#include <pluginsystem/pluginmanager.h>
+#include <core/plugins/pluginmanager.h>
 #include <utils/utils.h>
 
 #include <QContextMenuEvent>
