+++ src/gui/playlist/playlistwidget.cpp
@@ -30,7 +30,7 @@
 #include <core/library/librarymanager.h>
 #include <core/library/musiclibrary.h>
 #include <core/player/playermanager.h>
-#include <pluginsystem/pluginmanager.h>
+#include <core/plugins/pluginmanager.h>
 
 #include <QAction>
 #include <QActionGroup>
