+++ src/gui/widgets/splitterwidget.cpp
@@ -25,7 +25,7 @@
 
 #include <core/actions/actioncontainer.h>
 #include <core/actions/actionmanager.h>
-#include <pluginsystem/pluginmanager.h>
+#include <core/plugins/pluginmanager.h>
 #include <utils/enumhelper.h>
 
 #include <QJsonObject>
