--- src/libs/pluginsystem/CMakeLists.txt
@@ -1,9 +1,0 @@
-fooyin_add_library(PluginSystem
-    TYPE SHARED
-    PREFIX Yes
-    DEPENDS
-        Qt6::Core
-    SOURCES
-        plugininfo.cpp
-        pluginmanager.cpp
-)
