+++ src/plugins/filters/filterwidget.cpp
@@ -29,7 +29,7 @@
 #include <core/actions/actioncontainer.h>
 #include <core/library/musiclibrary.h>
 #include <core/player/playermanager.h>
-#include <pluginsystem/pluginmanager.h>
+#include <core/plugins/pluginmanager.h>
 #include <utils/enumhelper.h>
 
 #include <QAction>
